requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
attrs==25.1.0
backports.tarfile==1.2.0
bidict==0.23.1
blake3==1.0.4
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorful==0.5.6
compressed-tensors==0.8.1
depyf==0.18.0
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
docutils==0.21.2
einops==0.8.0
fastapi==0.115.7
filelock==3.17.0
fire==0.7.0
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
google-api-core==2.24.0
google-auth==2.38.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
interegular==0.3.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
lark==1.2.2
lazy_loader==0.4
lm-format-enforcer==0.10.9
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistral_common==1.5.2
mistral_inference==1.5.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
numpy==1.26.4
nvidia-ml-py==12.570.86
ollama==0.4.7
openai==1.60.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
partial-json-parser==0.2.1.1.post5
pillow==10.4.0
platformdirs==4.3.6
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.41.0
readme_renderer==44.0
redis==5.2.1
referencing==0.36.2
reflex==0.6.8
reflex-chakra==0.6.3
reflex-hosting-cli==0.1.33
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
sentencepiece==0.2.0
shellingham==1.5.4
simple-parsing==0.1.7
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.45.3
starlette-admin==0.14.1
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
tiktoken==0.7.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
twine==6.1.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
vllm==0.6.6.post1
watchfiles==1.0.4
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.29.post1
yarl==1.18.3
zipp==3.21.0
